Add submodule and logic refactor to handle the imported logic
