+++ .env.dist
@@ -8,3 +8,4 @@ VITE_ANALYTICS_APIKEY=<VITE_ANALYTICS_APIKEY>
 VITE_ANALYTICS_SENDERID=<VITE_ANALYTICS_SENDERID>
 VITE_ANALYTICS_APPID=<VITE_ANALYTICS_APPID>
 VITE_ANALYTICS_MEASUREMENTID=<VITE_ANALYTICS_MEASUREMENTID>
+VITE_CLIENTID=<VITE_CLIENTID>
