+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "learn-and-earn-submodule"]
+	path = learn-and-earn-submodule
+	url = git@github.com:impactMarket/learn-and-earn-submodule.git
