--- src/components/Breakpoints.tsx
@@ -1,8 +1,0 @@
-export const Breakpoints = {
-  xxsmall: "374px",
-  xsmall: "435px",
-  small: "576px",
-  medium: "768px",
-  large: "992px",
-  xlarge: "1200px",
-};
